fix(lint): fix lint + PR remarks
